# swagger.yaml
swagger: "2.0"
info:
  description: "Dashboard using different gaming apis(discord, twitch, battle.net and reddit)"
  version: "1.0.0"
  title: "Drainboard"
  termsOfService: "http://swagger.io/terms/"
host: "api.drainboard.tk"
...
schemes:
- "http"
paths:
  /auth/register:
    post:
      tags:
      - "auth"
      description: "Register an user with the single username/password protocol"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/RegisterLogin"
      responses:
        "422":
          description: "Unprocessable identity"
        "401":
          description: "Unauthorized"
  /auth/login:
    post:
      tags:
      - "auth"
      description: "Login an user with the username/password protocol"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/RegisterLogin"
      responses:
        "422":
          description: "Unprocessable identity"
        "401":
          description: "Unauthorized"
  /auth/bnet:
    get:
      tags:
      - "auth"
      description: "Redirect on the battle.net authorization page"
      produces:
      - "application/json"
      parameters:
      - name: "jwt"
        in: "query"
        description: "Token to find your authentified account"
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "302":
          description: "Redirect"
  /auth/bnet/return:
    get:
      tags:
      - "auth"
      description: "Redirect on the battle.net authorization page"
      produces:
      - "application/json"
      parameters:
      - name: "jwt"
        in: "query"
        description: "Token to find your authentified account"
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
        "401":
          description: "Unauthorized"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
  /auth/twitch:
    get:
      tags:
      - "auth"
      description: "Redirect on the Twitch authorization page"
      produces:
      - "application/json"
      parameters:
      - name: "jwt"
        in: "query"
        description: "Token to find your authentified account"
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "302":
          description: "Redirect"
  /auth/twitch/return:
    get:
      tags:
      - "auth"
      description: "Redirect on the battle.net authorization page"
      produces:
      - "application/json"
      parameters:
      - name: "jwt"
        in: "query"
        description: "Token to find your authentified account"
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
        "401":
          description: "Unauthorized"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
  /auth/discord:
    get:
      tags:
      - "auth"
      description: "Redirect on the Discord authorization page"
      produces:
      - "application/json"
      parameters:
      - name: "jwt"
        in: "query"
        description: "Token to find your authentified account"
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "302":
          description: "Redirect"
  /auth/discord/return:
    get:
      tags:
      - "auth"
      description: "Redirect on the Discord authorization page"
      produces:
      - "application/json"
      parameters:
      - name: "jwt"
        in: "query"
        description: "Token to find your authentified account"
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
        "401":
          description: "Unauthorized"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
  /auth/steam:
    get:
      tags:
      - "auth"
      description: "Redirect on the Steam authorization page"
      produces:
      - "application/json"
      parameters:
      - name: "jwt"
        in: "query"
        description: "Token to find your authentified account"
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "302":
          description: "Redirect"
  /auth/steam/return:
    get:
      tags:
      - "auth"
      description: "Redirect on the Steam authorization page"
      produces:
      - "application/json"
      parameters:
      - name: "jwt"
        in: "query"
        description: "Token to find your authentified account"
        required: true
        type: "string"
        collectionFormat: "multi"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
        "401":
          description: "Unauthorized"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
...
securityDefinitions:
  test:
    type: "oauth2"
    authorizationUrl: "http://Messagestore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:Messages: "modify Messages in your account"
      read:Messages: "read your Messages"
  jwt:
    type: "apiKey"
    name: "jwt"
    in: "header"
definitions:
  RegisterLogin:
    type: "object"
    properties:
      username:
        type: "string"
      password:
        type: "string"
...
  Message:
    type: "object"
    properties:
      message:
        type: "string"
...